chore: extract usefull utils from test
